requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.2
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==23.2.1
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.1
pytest-django==4.7.0
pytest-factoryboy==2.6.0
pytest-mock==3.12.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
